Minor refactoring + code cleanup
The point of this pull request is to help clean up some of the code.

I have also started refactoring things to an object. I believe this
will be more fitting as it moves forward, and we can never assume
that people aren't running more than one Express/Connect server
in the one node instance.

They might be running two different servers, and want to have two different
instances of spdy-referrer-push that do not clash.